+++ test/render-test.t/run.t
@@ -92,5 +92,6 @@
   Appending to a list literal works:pass pass pass pass 
 
   $ acutis dicts.acutis --data empty.json
-   success
+  success
+  success